refactor: excrements 관련 메소드 excrements 모듈로 이전
